Outsource util functions
